+++ tillerinobot/src/test/java/org/tillerino/ppaddict/util/LoggingUtilsTest.java
@@ -1,0 +1,14 @@
+package org.tillerino.ppaddict.util;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.Test;
+
+public class LoggingUtilsTest {
+
+	@Test
+	public void testGetRandomString() throws Exception {
+		assertThat(LoggingUtils.getRandomString(5)).matches("[A-Z]{5}");
+	}
+	
+}
